// src/types/authType.ts
interface Subscription {
  id: string;
}

interface ResponseData {
  success: boolean;
  data: {
    message: string;
    subscription: Subscription;
  };
}

export interface ApiResponseCancelSubscription {
  data: ResponseData;
  success: boolean;
}

interface InnerData {
  message: string;
}

interface Data {
  success: boolean;
  data: InnerData;
}

export interface PauseandResumeApiResponse {
  data: Data;
  success: boolean;
}